fixed 3 more warnings

+++ DebugView++Lib/FileWriter.cpp
@@ -32,7 +32,7 @@ void FileWriter::Process()
 {
 	//todo: we need locking on Logfile, think of ClearLog() 
 	// also, reading the .dblog file does not work correctly
-	int writeIndex = 0;
+	size_t writeIndex = 0;
 	for(;;)
 	{
 		while (m_logfile.Count() > writeIndex)
